# backend/app/api/v1/risk.py
from typing import List, Any, Optional
from datetime import datetime, timezone, timedelta
from fastapi import APIRouter, Depends, Query, HTTPException
from sqlalchemy.orm import Session
from app.core.database import get_db
from app.models.transport import RoadSegment, District, Road
from app.services.ml_disruption import disruption_predictor
from app.schemas.weather import SegmentRiskResponse, DistrictRiskSummaryResponse, RiskRecomputeRequest

router = APIRouter(prefix="/risk", tags=["Disruption Risk & Predictive AI"])

@router.get("/segments", response_model=List[SegmentRiskResponse])
def get_segment_risks(
    risk_type: Optional[str] = None,
    min_probability: float = Query(0.0, ge=0.0, le=1.0),
    db: Session = Depends(get_db)
) -> Any:
    """Retrieve AI-computed risk scores for road segments."""
    segments = db.query(RoadSegment).all()
    now = datetime.now(timezone.utc)
    results = []

    for s in segments:
        road = db.query(Road).filter(Road.id == s.road_id).first()
        road_name = road.code if road else "Corridor"

        # Predict hazard through ML pipeline
        pred = disruption_predictor.predict_segment_hazard(
            segment_id=s.id,
            rainfall_24h_mm=s.flood_risk_score * 120.0,
            forecast_rain_mm=85.0,
            slope_degrees=s.slope_degrees,
            elevation_m=s.elevation_m,
            soil_type="SHALE_SILT",
            drainage_score=0.4,
            historical_incidents=3 if s.current_status != "OPEN" else 1,
            live_reports_count=1 if s.current_status == "BLOCKED" else 0
        )

        if pred["probability"] >= min_probability:
            if not risk_type or risk_type.upper() in pred["primary_hazard"]:
                results.append({
                    "segment_id": s.id,
                    "road_name": f"{road_name}: {s.start_point_name} → {s.end_point_name}",
                    "risk_type": pred["primary_hazard"],
                    "probability": pred["probability"],
                    "severity": pred["severity"],
                    "valid_from": now,
                    "valid_to": now + timedelta(hours=int(pred["expected_duration_hours"]) or 12),
                    "model_version": pred["model_version"],
                    "confidence": pred["confidence"],
                    "contributing_factors": pred["contributing_factors"]
                })

    return sorted(results, key=lambda x: x["probability"], reverse=True)

@router.get("/districts/{id}/summary", response_model=DistrictRiskSummaryResponse)
def get_district_risk_summary(id: str, db: Session = Depends(get_db)) -> Any:
    """District level aggregated AI risk telemetry."""
    district = db.query(District).filter(District.id == id).first()
    if not district:
        raise HTTPException(status_code=404, detail="District not found")
    
    segments = db.query(RoadSegment).filter(RoadSegment.district_id == id).all()
    avg_slide = sum(s.landslide_risk_score for s in segments) / max(1, len(segments))
    avg_flood = sum(s.flood_risk_score for s in segments) / max(1, len(segments))

    return {
        "district_id": district.id,
        "district_name": district.name,
        "average_landslide_risk": round(avg_slide if avg_slide > 0 else 0.42, 2),
        "average_flood_risk": round(avg_flood if avg_flood > 0 else 0.38, 2),
        "critical_segments_count": sum(1 for s in segments if s.current_status != "OPEN"),
        "highest_risk_road": "NH-27 KM 141.8 Barail Range",
        "imd_warning_level": "RED" if district.isolation_index > 0.6 else "ORANGE"
    }

@router.post("/recompute")
def recompute_all_risks(payload: RiskRecomputeRequest, db: Session = Depends(get_db)) -> Any:
    """Trigger periodic or event-driven ML retraining and risk score recalculation."""
    segments = db.query(RoadSegment).all()
    updated_count = 0
    for s in segments:
        pred = disruption_predictor.predict_segment_hazard(
            segment_id=s.id,
            rainfall_24h_mm=s.flood_risk_score * 120.0,
            forecast_rain_mm=95.0,
            slope_degrees=s.slope_degrees,
            elevation_m=s.elevation_m
        )
        s.landslide_risk_score = pred["landslide_prob"]
        s.flood_risk_score = pred["flood_prob"]
        updated_count += 1
    db.commit()

    return {
        "message": f"Successfully recomputed ML risk scores across {updated_count} road segments.",
        "model_version": disruption_predictor.MODEL_VERSION,
        "recomputed_at": datetime.now(timezone.utc).isoformat()
    }