# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NER-LogiSense Intelligence Platform",
    "description": "\n# NER-LogiSense Backend API\n### AI-powered Smart Logistics and Accessibility Intelligence Platform for India's North Eastern Region (NER)\n\n**Core Capabilities**:\n- Real-time road, bridge, and transport accessibility monitoring\n- Landslide and flood disruption prediction using environmental feature models\n- AI-based multi-criteria route optimization ($Cost = \\alpha \\cdot T + \\beta \\cdot R + \\gamma \\cdot C$)\n- High-throughput GPS tracking of vehicles carrying medical oxygen, cryo-vaccines, foodgrains, and relief supplies\n- Offline-first field hazard reporting with idempotency and conflict resolution\n- Multilingual automated alerts across 9 regional languages\n- WebSocket telemetry streaming\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/risk/segments": {
      "get": {
        "tags": [
          "Disruption Risk & Predictive AI"
        ],
        "summary": "Get Segment Risks",
        "description": "Retrieve AI-computed risk scores for road segments.",
        "operationId": "get_segment_risks_api_v1_risk_segments_get",
        "parameters": [
          {
            "name": "risk_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Type"
            }
          },
          {
            "name": "min_probability",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.0,
              "title": "Min Probability"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SegmentRiskResponse"
                  },
                  "title": "Response Get Segment Risks Api V1 Risk Segments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/districts/{id}/summary": {
      "get": {
        "tags": [
          "Disruption Risk & Predictive AI"
        ],
        "summary": "Get District Risk Summary",
        "description": "District level aggregated AI risk telemetry.",
        "operationId": "get_district_risk_summary_api_v1_risk_districts__id__summary_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictRiskSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/risk/recompute": {
      "post": {
        "tags": [
          "Disruption Risk & Predictive AI"
        ],
        "summary": "Recompute All Risks",
        "description": "Trigger periodic or event-driven ML retraining and risk score recalculation.",
        "operationId": "recompute_all_risks_api_v1_risk_recompute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskRecomputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Recompute All Risks Api V1 Risk Recompute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync/queue": {
      "post": {
        "tags": [
          "Offline-First Synchronization"
        ],
        "summary": "Sync Offline Queue",
        "description": "Synchronize batched offline queue items:\n- Guarantees idempotency via client_uuid\n- Implements conflict resolution\n- Preserves causal order",
        "operationId": "sync_offline_queue_api_v1_sync_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/delta": {
      "get": {
        "tags": [
          "Offline-First Synchronization"
        ],
        "summary": "Get Delta Sync",
        "description": "Provides delta synchronization using updated_at cursor.\nAllows mobile clients to fetch only entities modified since last synchronization.",
        "operationId": "get_delta_sync_api_v1_sync_delta_get",
        "parameters": [
          {
            "name": "last_synced_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO timestamp cursor from previous sync",
              "title": "Last Synced At"
            },
            "description": "ISO timestamp cursor from previous sync"
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeltaSyncResponse": {
        "properties": {
          "server_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Server Timestamp"
          },
          "updated_roads": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Updated Roads"
          },
          "updated_segments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Updated Segments"
          },
          "active_incidents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Active Incidents"
          },
          "active_alerts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Active Alerts"
          },
          "active_convoys": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Active Convoys"
          }
        },
        "type": "object",
        "required": [
          "server_timestamp",
          "updated_roads",
          "updated_segments",
          "active_incidents",
          "active_alerts",
          "active_convoys"
        ],
        "title": "DeltaSyncResponse"
      },
...
      "DistrictRiskSummaryResponse": {
        "properties": {
          "district_id": {
            "type": "string",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "average_landslide_risk": {
            "type": "number",
            "title": "Average Landslide Risk"
          },
          "average_flood_risk": {
            "type": "number",
            "title": "Average Flood Risk"
          },
          "critical_segments_count": {
            "type": "integer",
            "title": "Critical Segments Count"
          },
          "highest_risk_road": {
            "type": "string",
            "title": "Highest Risk Road"
          },
          "imd_warning_level": {
            "type": "string",
            "title": "Imd Warning Level"
          }
        },
        "type": "object",
        "required": [
          "district_id",
          "district_name",
          "average_landslide_risk",
          "average_flood_risk",
          "critical_segments_count",
          "highest_risk_road",
          "imd_warning_level"
        ],
        "title": "DistrictRiskSummaryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskRecomputeRequest": {
        "properties": {
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "force_all": {
            "type": "boolean",
            "title": "Force All",
            "default": false
          }
        },
        "type": "object",
        "title": "RiskRecomputeRequest"
      },
...
      "SegmentRiskResponse": {
        "properties": {
          "segment_id": {
            "type": "string",
            "title": "Segment Id"
          },
          "road_name": {
            "type": "string",
            "title": "Road Name"
          },
          "risk_type": {
            "type": "string",
            "title": "Risk Type"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "valid_to": {
            "type": "string",
            "format": "date-time",
            "title": "Valid To"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "contributing_factors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributing Factors"
          }
        },
        "type": "object",
        "required": [
          "segment_id",
          "road_name",
          "risk_type",
          "probability",
          "severity",
          "valid_from",
          "valid_to",
          "model_version",
          "confidence"
        ],
        "title": "SegmentRiskResponse"
      },
...
      "SyncBatchRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "sync_session_id": {
            "type": "string",
            "title": "Sync Session Id"
          },
          "queue_items": {
            "items": {
              "$ref": "#/components/schemas/SyncQueueItem"
            },
            "type": "array",
            "title": "Queue Items"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "sync_session_id",
          "queue_items"
        ],
        "title": "SyncBatchRequest"
      },
      "SyncBatchResponse": {
        "properties": {
          "sync_session_id": {
            "type": "string",
            "title": "Sync Session Id"
          },
          "processed_count": {
            "type": "integer",
            "title": "Processed Count"
          },
          "success_count": {
            "type": "integer",
            "title": "Success Count"
          },
          "conflict_count": {
            "type": "integer",
            "title": "Conflict Count"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncItemResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "sync_session_id",
          "processed_count",
          "success_count",
          "conflict_count",
          "results"
        ],
        "title": "SyncBatchResponse"
      },
      "SyncItemResult": {
        "properties": {
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "server_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id"
          },
          "conflict_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Details"
          },
          "processed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Processed At"
          }
        },
        "type": "object",
        "required": [
          "client_uuid",
          "status",
          "processed_at"
        ],
        "title": "SyncItemResult"
      },
      "SyncQueueItem": {
        "properties": {
          "queue_id": {
            "type": "string",
            "title": "Queue Id"
          },
          "client_uuid": {
            "type": "string",
            "title": "Client Uuid"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "client_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Client Timestamp"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "queue_id",
          "client_uuid",
          "operation",
          "payload",
          "client_timestamp",
          "device_id"
        ],
        "title": "SyncQueueItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}